Automatically backed up by Learn

+++ lib/create.sql
@@ -10,4 +10,4 @@ title TEXT, category TEXT, funding_goal REAL, start_date TEXT, end_date TEXT);
 
 CREATE TABLE projects (
 id INTEGER PRIMARY KEY,
-amount INTEGER, user_id INTEGER, project_id INTEGER);
+amount REAL, user_id INTEGER, project_id INTEGER);
